the constants.js was moved to the root client's folder, the refreshHash and the switchFormFields were renamed to the setHash and the clearFormFields
