+++ src/client/utils/clearFormFields.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign */
-export function refreshFormFields(form) {
+export function clearFormFields(form) {
   Array.from(form.elements).forEach((item) => {
     item.value = '';
   });
