+++ src/client/utils/index.js
@@ -1,6 +1,5 @@
 export { getElementById, getHash, getPropsFormFields, getUrl } from './getters';
 export { EventEmitter } from './eventEmitter';
-export { URLS, EVENTS, BUTTONS, FORMS } from './constants';
 export { createList, createDescription } from './creators';
-export { refreshFormFields } from './refreshFormFields';
-export { switchHash } from './switchHash';
+export { clearFormFields } from './clearFormFields';
+export { setHash } from './setHash';
